Update 3-account-linking-md
There is an additional inital prompt in the latest version of the ASK CLI that now appears which is not noted here: `? Allow users to enable skill without account linking:  Yes` - if this is set to No, simulation fails. Also, clarified that the Client ID is the Salesforce Consumer Key in this example.